+++ src/Bedrock/Application/Kernel.php
@@ -13,7 +13,7 @@ class Kernel
     /**
      * Framework version
      */
-    const VERSION = '2.27.0';
+    const VERSION = '2.28.0';
     
     /**
      * App bootstrap object if exists