+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+VERSION 2.28.0
+--------------
+Release Date: ?
+
 VERSION 2.27.0
 --------------
 Release Date: 2018-05-20